Rework hybrid height load test.
