+++ iris_grib/tests/unit/load_convert/test_vertical_coords.py
@@ -29,8 +29,6 @@ import iris_grib.tests as tests
 from copy import deepcopy
 import mock
 
-from iris import load_cube
-from iris.aux_factory import HybridHeightFactory
 from iris.coords import DimCoord
 from iris.exceptions import TranslationError
 
@@ -173,20 +171,6 @@ class Test(tests.IrisGribTest):
         expected = deepcopy(self.metadata)
         expected['aux_coords_and_dims'].append((coord, None))
         self.assertEqual(metadata, expected)
-
-
-class TestRealData(tests.IrisGribTest):
-    def test_load_hybrid_height(self):
-        """
-        Temporary test for development.  Replace with proper test  or tidy up
-        for public consumption before making final PR.
-        """
-        import os.path
-        here_dirpath = os.path.dirname(__file__)
-        filepath = here_dirpath + '/../../../test_data/'
-        fname = 'faked_sample_hh_grib_data.grib2'
-        cube = load_cube(filepath + fname, 'air_temperature')
-        self.assertIsInstance(cube.aux_factories[0], HybridHeightFactory)
 
 
 if __name__ == '__main__':
